+++ platforms/linux/fpm.sh
@@ -14,6 +14,13 @@ find ${WORKDIR} -maxdepth 1 -type f -not -name sno_cli -exec chmod -x {} \;
 # symlink executable
 ln -sf /opt/sno/sno_cli ${WORKDIR}/usr/bin/sno
 
+OPTS=
+if [ "$TYPE" = "deb" ]; then
+    OPTS+="--depends openssh-client"
+elif [ "$TYPE" = "rpm" ]; then
+    OPTS+="--depends openssh-clients"
+fi
+
 # build package
 fpm \
     --verbose \
@@ -23,7 +30,9 @@ fpm \
     --name sno \
     --version "${VERSION}" \
     --url "https://sno.earth" \
+    --license "GPLv2" \
     --architecture amd64 \
     --package /src/dist/ \
     --force \
+    $OPTS \
     .